transcribe_audio/requirements.txt
openai-whisper
ffmpeg-python
pydub 